fix answer.php

+++ answer.php
@@ -1,52 +1,54 @@
 <html>
 <head>
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
+          integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
+          crossorigin="anonymous">
 </head>
 <body>
 <?php
-$servername = "localhost";
-$username = "kali";
-$password = "kali";
-$dbname = "game";
+$serverName = "localhost";
+$connectionOptions = array(
+    "Database" => "game",
+    "UID" => "kali",
+    "PWD" => "kali"
+);
 
-// Create connection
-$conn = mysqli_connect($servername, $username, $password, $dbname);
-// Check connection
-if (!$conn) {
-  die("Connection failed: " . mysqli_connect_error());
+$conn = sqlsrv_connect($serverName, $connectionOptions);
+if ($conn === false) {
+    die("Connection failed: " . print_r(sqlsrv_errors(), true));
 }
 
-$sql = "SELECT hints from challenge ORDER BY chalID DESC LIMIT 1";
-mysqli_query($conn, $sql);
-$result = mysqli_query($conn, $sql);
+$sql = "SELECT hints FROM challenge ORDER BY chalID DESC OFFSET 0 ROWS FETCH NEXT 1 ROWS ONLY";
+$result = sqlsrv_query($conn, $sql);
 
-if (mysqli_num_rows($result) > 0) {
-  // output data of each row
-  while($row = mysqli_fetch_assoc($result)) {
-    echo "hints: " . $row["hints"];
-  }
+if ($result !== false) {
+    if (sqlsrv_has_rows($result)) {
+        while ($row = sqlsrv_fetch_array($result, SQLSRV_FETCH_ASSOC)) {
+            echo "hints: " . $row["hints"];
+        }
+    } else {
+        echo "0 results";
+    }
 } else {
-  echo "0 results";
+    echo "Error executing query: " . print_r(sqlsrv_errors(), true);
 }
 
-mysqli_close($conn);
+sqlsrv_close($conn);
 ?>
 <?php
 if (isset($_POST["btn_submit"])) {
-$answer = $_POST["answer"];
-		if ($answer == "") {
-			echo "Answer is empty!";
-		}else{
-			$myfile = fopen("challenge/$answer.txt", "r");
-			if($myfile){
-			echo "<br>Correct<br>";
-			echo fread($myfile,filesize("challenge/$answer.txt"));
-			fclose($myfile);
-			}
-			else{
-			echo "<br>Incorrect";
-			}
-		}
+    $answer = $_POST["answer"];
+    if ($answer == "") {
+        echo "Answer is empty!";
+    } else {
+        $file = "challenge/" . $answer . ".txt";
+        if (file_exists($file)) {
+            echo "<br>Correct<br>";
+            echo file_get_contents($file);
+        } else {
+            echo "<br>Incorrect";
+        }
+    }
 }
 ?>
 <form method="POST" action="answer.php">
